make postgres happy

+++ ephios/core/models/users.py
@@ -159,7 +159,7 @@ class UserProfile(guardian.mixins.GuardianUserMixin, PermissionsMixin, AbstractB
                         (
                             F("shift__end_time") - F("shift__start_time")
                         ),  # calculate length of shift in μs
-                        output_field=models.DecimalField(),
+                        output_field=models.IntegerField(),
                     )
                     / 1000000.0
                     / 3600.0,  # convert microseconds to seconds to hours
